Reduced the amount of memory used by lightmaps in the renderer of the OXR version by using 16-bit values instead of 32-bit ones - the extra space was never needed to begin with. The LightmapTexture struct has no longer any default initializers; dynamic creation using the new() operator will now initialize by way of the { } syntax.

+++ id1/d_lists.h
@@ -181,7 +181,7 @@ struct dlists_t
 	std::vector<unsigned char> colored_indices8;
 	std::vector<uint16_t> colored_indices16;
 	std::vector<uint32_t> colored_indices32;
-	std::vector<unsigned> lightmap_texels;
+	std::vector<uint16_t> lightmap_texels;
 };
 
 extern dlists_t d_lists;
